Started spherical plot3d mode implementation(not working)

+++ plot.py
@@ -23,19 +23,31 @@ set zlabel "z"
 set datafile missing "NaN"
 splot "-" title "Function" with pm3d
 '''
-def plot3d(func, var1_range, var2_range, scale_x=10, scale_y=10):
+def plot3d(func, var1_range, var2_range, mode='cartesian', scale_x=10, scale_y=10):
+    if type(func).__name__ in ('int', 'float'):
+        func = sympy.Number(func)
+
     var1 = sympy.Symbol(var1_range[0])
     var2 = sympy.Symbol(var2_range[0])
 
-    step_x = (var1_range[-1] - var1_range[1])/scale_x
-    step_y = (var2_range[-1] - var2_range[1])/scale_y
+    if mode == 'cartesian':
+        step_x = (var1_range[-1] - var1_range[1])/scale_x
+        step_y = (var2_range[-1] - var2_range[1])/scale_y
+    elif mode == 'spherical':
+        step_x = ((sympy.sin(var1_range[-1])*sympy.cos(var2_range[-1]) 
+                   - sympy.sin(var1_range[1])*sympy.cos(var2_range[1])) / scale_x)
+        step_y = ((sympy.sin(var1_range[-1])*sympy.sin(var2_range[-1]) 
+                   - sympy.sin(var1_range[1])*sympy.sin(var2_range[1])) / scale_y)
 
     output = ''
     for i in range(scale_x):
         var1_val = float(i*step_x)
         for j in range(scale_y):
             var2_val = float(j*step_y)
-            output += str(var1_val) + ' ' + str(var2_val) + ' ' + str(func.subs(var2, var2_val).subs(var1, var1_val)) + '\n'
+            try:
+                output += str(var1_val) + ' ' + str(var2_val) + ' ' + str(func.subs(var2, var2_val).subs(var1, var1_val)) + '\n'
+            except AttributeError:
+                return 'undef'
         output += '\n'
     
     return output
@@ -44,18 +56,19 @@ x = sympy.Symbol('x')
 y = sympy.Symbol('y')
 
 #gp_input = open('test.gnuplot', 'r').read()
-gp_input = plot3d(x*y, ['x',0,10], ['y',0,10], 5, 5)
+gp_input = plot3d(1, ['x',0,10], ['y',0,10], 'spherical', 5, 5)
 
-gp_filename = 'temp.gnuplot'
-gp_fd = open(gp_filename, 'w')
-gp_fd.write(gp_settings)
-gp_fd.write(gp_input)
-gp_fd.close()
-
-p = subprocess.Popen(['gnuplot', '-p', gp_filename])
-p.wait()
-
-try:
-    os.remove(gp_filename)
-except OSError as (errno, strerror):
-    print(gp_filename + ': ' + strerror)
+if gp_input != 'undef':
+    gp_filename = 'temp.gnuplot'
+    gp_fd = open(gp_filename, 'w')
+    gp_fd.write(gp_settings)
+    gp_fd.write(gp_input)
+    gp_fd.close()
+    
+    p = subprocess.Popen(['gnuplot', '-p', gp_filename])
+    p.wait()
+    
+    try:
+        os.remove(gp_filename)
+    except OSError as (errno, strerror):
+        print(gp_filename + ': ' + strerror)
